correction univ

+++ club-app-front/src/app/university/list-university/list-university.component.ts
@@ -1,62 +1,88 @@
 import { Component, OnInit } from '@angular/core';
-import { Router } from '@angular/router';
-import { UniversityServiceService } from 'src/app/shared/university-service.service';
 import { University } from '../university';
+import { UniversityServiceService } from 'src/app/shared/university-service.service';
+import { Router } from '@angular/router';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-list-university',
   templateUrl: './list-university.component.html',
   styleUrls: ['./list-university.component.css']
 })
-export class listUniversityComponent implements OnInit {
+export class ListUniversityComponent implements OnInit {
+  universities:University[];
+   sear:String;
 
-  
-  universities: University[];
+  // universities: University[];
   constructor(private UniversityServiceservice : UniversityServiceService,
-    private router: Router ) { }
+    private router: Router) { }
 
   ngOnInit(): void {
-    this.universities = [];
 
-    this.getUniversities();
+     this.sear=""
+     this.universities = [];
+
+     this.getUniversities();
 
     
-  }
+   }
 
-  private getUniversities(){
-    this.UniversityServiceservice.getallUniversities().subscribe(data=> {
-      console.log(data);
-      this.universities = data;
-    })
-  }
+   private getUniversities(){
+     this.UniversityServiceservice.getallUniversities().subscribe(data=> {
+       console.log(data);
+       this.universities = data;
+     })
+   }
 
-  UpdateUniversity(UnivId:number){
+
+   UpdateUniversity(UnivId:number){
     this.router.navigate(['UpdateUniversity',UnivId]);
-
+    console.log('error');
   }
-  // deleteUniversite(UnivId:number){
-  //   console.log(UnivId)
-  //   this.UniversityServiceservice.deleteUniversite(UnivId).subscribe (data =>{
-  //       console.log(data);
-  //       this.getUniversities();
-  //   })
+
+
+  
   
     deleteUniversite(idUniv: number){
-      console.log(idUniv),
-      this.UniversityServiceservice.deleteUniversite(idUniv).subscribe (data =>{
+       console.log(idUniv),
+      
+     Swal.fire({
+       title: 'Are you sure?',
+       text: "You won't be able to revert this!",
+       icon: 'warning',
+       showCancelButton: true,
+       confirmButtonColor: '#3085d6',
+       cancelButtonColor: '#d33',
+       confirmButtonText: 'Yes, delete it!'
+     }).then((result) => {
+       if (result.isConfirmed) {
+         this.UniversityServiceservice.deleteUniversite(idUniv).subscribe (data =>{
            console.log(data);
-           this.getUniversities(); 
-        
-       window.location.reload();
+          // this.getUniversities();      
     })
-  }
+         Swal.fire(
+           'Deleted!',
+           'Your file has been deleted.',
+           'success',
+          
+         )
+         window.location.reload();
+       }
+     })
+    
+   }
 
 
   
-  detailsUniversite(UnivId:number){
-    this.router.navigate(['DetailsUniversity',UnivId]);
+   detailsUniversite(UnivId:number){
+     this.router.navigate(['DetailsUniversity',UnivId]);
 
-  }
+   }
 
   
 }
+ 
+  
+
+  
+
